Simplified client application versioning at startup
There was always a requirement that the client's version match the xcodeproj version exactly.  This wouldn't be possible if the patch version was hardcoded into the client app's code and the xcodeproj patch version was incremented automatically by a tool.  Thus, now the version is, by default, pulled from the xcodeproj and the user does not need to specify it during startup unless they want to do so.