Created a macro @ViewModelImpl that augments View Models
The proper way to specify a ViewModel is now:

```swift
@ViewModelImpl struct MyViewModel {}
```

There's no need (although it is permitted) to specify conformance to ViewModel; it's given automatically. This macro also injects the property name lookup info needed to bind the property names to the @LocalizedString property wrappers.
This fixes an issue where you could not have ViewModels be properties of other view models (although this is still a problem if any substitutable property wrappers are used ... future).

All Errors now conform to CustomDebugStringConvertable so that they print out nice messages with String(describing:) and in the debug logs.